changes in Team Controller to manage serie when team is deleted

+++ SB_backend/Controllers/PlayerController.cs
@@ -85,7 +85,8 @@ namespace SB_backend.Controllers
 
             if (current_player != null)
             {
-                System.IO.File.Delete(Path.Combine(Directory.GetCurrentDirectory(), current_player.ImgPath));
+                if(current_player.ImgPath != null)
+                    System.IO.File.Delete(Path.Combine(Directory.GetCurrentDirectory(), current_player.ImgPath));
                 this.SaveFile(player);
                 player.Id = current_player.Id;
                 _plrep.UpdatePlayer(player, positions);
